Move MemoryPersister impl to own file
